Passing norminette for this auto selecting algorithm

+++ src/sorting/algos/move_all.c
@@ -6,7 +6,7 @@
 /*   By: ehosta <ehosta@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/27 09:23:10 by ehosta            #+#    #+#             */
-/*   Updated: 2025/01/27 11:35:35 by ehosta           ###   ########.fr       */
+/*   Updated: 2025/01/27 14:45:38 by ehosta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@ int	move_reverse_all(t_env *env, int emulation)
 {
 	int	i;
 	int	original_size;
-	int		moves;
+	int	moves;
 
 	i = -1;
 	moves = 0;
 	original_size = env->a_size;
-	while ((int)++i < original_size)
+	while (++i < original_size)
 	{
 		if (env->stack_a[env->a_head] > 2)
 			push(env, 'a', !emulation);
@@ -36,14 +36,13 @@ int	everything_ascending(t_env *env, int emulation)
 {
 	int	i;
 	int	original_size;
-	int		moves;
+	int	moves;
 
 	i = -1;
 	moves = 0;
 	original_size = env->a_size;
-	while ((int)++i < original_size)
+	while (++i < original_size)
 	{
-		
 		if (env->stack_a[env->a_head] > 2)
 			push(env, 'a', !emulation);
 		else
